fix: change logic back to normal
